@@ -2,10 +2,10 @@
 
   <PropertyGroup>
     <OutputType>Library</OutputType>
-    <TargetFramework>net6.0-windows</TargetFramework>
+    <TargetFramework>net7.0-windows</TargetFramework>
     <UseWindowsForms>true</UseWindowsForms>
-    <AssemblyVersion>2022.11.10.1021</AssemblyVersion>
-    <FileVersion>2022.11.10.1021</FileVersion>
+    <AssemblyVersion>2022.11.10.1150</AssemblyVersion>
+    <FileVersion>2022.11.10.1150</FileVersion>
   </PropertyGroup>
 
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
