@@ -10,7 +10,7 @@ internal static class Version
 
     static public string History =
         "History" +
-        "\r\n ver4.840(2022/11/10)  Fixed a bug that occurred when starting 'Diffraction Simulator' (see https://github.com/seto77/ReciPro/issues/14)." +
+        "\r\n ver4.840(2022/11/10)  Fixed a bug that occurred when starting 'Diffraction Simulator' (see https://github.com/seto77/ReciPro/issues/16)." +
         "\r\n ver4.839(2022/11/07)  Added a function to simulate X-ray precession camara." +
         "\r\n ver4.838(2022/10/21)  Improved compatibility of importing CIF files." +
         "\r\n ver4.837(2022/10/20)  Added a function to output superstructure." +
